Successful Render
It all must be contained within a single svg, otherwise they dont overlap as easily. That being the case, I used a <g> tag run the first for loop.

Glyphs and links both occur within a second loop– through the path itmes. Origionally, I wanted to filter this loop in Links.vue. however, if its required for glyphs, it is not wasted.

Paths, Links, and Glyphs.vue templates are now much easier to read.