Z-index Render Conflict
You will notice that the green path is ontop of a blue dot. Previously all dots were rendered after paths. However, but combining the for loops that is not necessarily true. The easiest solution is to remove in-path links or assign a z-index to the svgs some how...

cleaned scrap work and formatted templates a bit.