Demonestrates Functionality
This is either the dumbest or smartest code I have ever written. It loops through id's and renders an ailias later with a new z-index.

While it is condensed and guarantees glyphs are rendered on top, It does not guarentee that links will be.

I will try making an alias template to make this cleaner,  but its begining to get a bit round about.

It might simply be cleaner, faster, and more intuitive to leave as explicit loops for each. hence the new branch.