Cleaned but wont pull
While it does function, it failed to achieve my intentions– reducing for loops and improving clarity. (condensing is inferior to these)

This unfortunately doesnt impove clarity, it hides the loops and muddies the final HTML.

I may as well have created a component to outsource the for loops prior to branching. That would have been more direct than use at least.