Children X and dX function properly
Children cause some parent branches to move when active. Thus, it was a prerequisite to implement dx functionality. Cascade of dx parameter changes was made.

For proper child positioning, the x list must be summed. Such functionality was added to Paths.vue{ getPath() } and was already functional in DisplayMixins.

Next up:
- cleanup parameter ordering in DisplayMixins.js
-Toggle show children
-still uncertain as to whether x-overrides should be a dx of xConst...

Overview of Diffs:
/DisplayMixin.js
- Added dx parameter and value to getXDisp.

/index.js
- Added child branch P1.1
- Added branch next to child to test dx.

/Paths.vue
- xConst now sum of parent and children.

/Links.vue & /Glyphs.vue
- updated new dx parameter in getXDisp