Prep for conditional rendering
Essentials for conditional rendering of children have been put into place.

Branches to be displayed will be added to $store.state.show
While added and removing data is roughly implemented, it is not active yet. That will be the following commit.

double click parent branch to toggle children. Activation is functional, but conditional rendering is not. That is next commit.

Added glow to show branches hiding children.