Failure – will abandon
its just not a good method no point in pushing the issue...

ISSUES:
While it is sort of close to functional, its fails some important aspects:
• If child has different x-list, it wont be found and collapsed, ever.
• Decendants may recieve too much/many dx changes (if using dXUpdate)
• If using dxReset, lots of issues... like thing that are shifted by 2 events would be reset to 0 rather than shifted by just one.
• etc.

NOTE:
recursion is probably better as it garuntees only registered children collapse. It is very possible that register children may not share X list or X list may be shared and not registered as a collapsable child. In some ways these would likely be considered bad form, however, if used  the behaviour would be unexpected.

Is parent-child binding an alternative?
Probably not. It doesnt make getting deeper descendants any easier.