Successful render- failed experiment
This method requires that displayed paths be tracked, namely in show.

because I can not circumvent this the recursive element in the function does pretty much nothing. there may be some way to update on show witch changes in parent, but dont know a good way.

As such, I have re added the recursive toggle method.