Junction Point
I wanted to add the necessary data to this file to keep it localized and easier to access. However, mixins are not meant share data, just funtionality like  methods.
As a result they are not terribly reactive. The two solution I wish to try are making it a global mixin and adding it back to vuex perhaps as a module.